DCS-97 splitting out incident handlers from router definition
